Ignore words that have more phonemes than graphemes
CTC is an alignment method where the length of the target sequence
is at most the length of the input sequence. In our case, specifically,
the number of phonemes cannot exceed the number of graphemes in a given
string. This is the case for about 2k of the 110k words.

This commit addresses a bug that prevents correct training on the GPU.
cpu_ctc returns 0 in the above case (more phonemes than graphemes), but
gpu_ctc returns an extremely large number (e.g. overflow). (As an aside,
I believe the behavior of gpu_ctc should be expected, but it's annoying
it failed silently).

Discovered by saving the predictions, targets, inputs, and sizes for
error cases, and running gpu_ctc and cpu_ctc.