[GPU] Do not cast model to double and back every epoch
Cuda corresponds to float, while the default Tensor type is double.
It appears this back and forth conversion causes numerical instability.
I originally had it this way so that I could train on GPU and easily load
on CPU, but whatever.

Discovered when loss on GPU plateaued starting from the second epoch.