Handle last minibatch of epoch
Last minibatch may not be equal to opt.minibatch if the number of points
is not a multiple of opt.minibatch.

Since the batchsize is now non-deterministic, we can't use it as a parameter
when constructing our network. The last View module to reshape the output
doesn't work. Therefore, we reshape the activations from our network.
Note, we also have to reshape the grad_output before calling backward.

(Actually, batchsize doesn't even need to be passed to g2p_model anymore.)

There is probably a cleaner way of doing this.